+++ modules/matrix-bridge/test/gitter-bridge-test.js
@@ -27,6 +27,7 @@ describe('gitter-bridge', () => {
       createRoom: sinon.spy(() => ({
         room_id: `!${fixtureLoader.generateGithubId()}:localhost`
       })),
+      createAlias: sinon.spy(),
       setDisplayName: sinon.spy(),
       uploadContent: sinon.spy(),
       setAvatarUrl: sinon.spy()
@@ -145,12 +146,10 @@ describe('gitter-bridge', () => {
           strategy
         );
 
-        await store.storeBridgedRoom(
-          fixture.troupe1.id,
-          `!${fixtureLoader.generateGithubId()}:localhost`
-        );
+        const matrixRoomId = `!${fixtureLoader.generateGithubId()}:localhost`;
+        await store.storeBridgedRoom(fixture.troupe1.id, matrixRoomId);
         const parentMessageEventId = `$${fixtureLoader.generateGithubId()}:localhost`;
-        await store.storeBridgedMessage(fixture.message1.id, parentMessageEventId);
+        await store.storeBridgedMessage(fixture.message1.id, matrixRoomId, parentMessageEventId);
 
         await gitterBridge.onDataChange({
           url: `/rooms/${fixture.troupe1.id}/chatMessages`,
@@ -183,14 +182,12 @@ describe('gitter-bridge', () => {
           strategy
         );
 
-        await store.storeBridgedRoom(
-          fixture.troupe1.id,
-          `!${fixtureLoader.generateGithubId()}:localhost`
-        );
+        const matrixRoomId = `!${fixtureLoader.generateGithubId()}:localhost`;
+        await store.storeBridgedRoom(fixture.troupe1.id, matrixRoomId);
         // We purposely do not associate the bridged message. We are testing that the
         // message is ignored if the parent message event is not in the database.
         //const parentMessageEventId = `$${fixtureLoader.generateGithubId()}:localhost`;
-        //await store.storeBridgedMessage(fixture.message1.id, parentMessageEventId);
+        //await store.storeBridgedMessage(fixture.message1.id, matrixRoomId, parentMessageEventId);
 
         await gitterBridge.onDataChange({
           url: `/rooms/${fixture.troupe1.id}/chatMessages`,
@@ -286,8 +283,9 @@ describe('gitter-bridge', () => {
         const strategy = new restSerializer.ChatStrategy();
         const serializedMessage = await restSerializer.serializeObject(fixture.message1, strategy);
 
+        const matrixRoomId = `!${fixtureLoader.generateGithubId()}:localhost`;
         const matrixMessageEventId = `$${fixtureLoader.generateGithubId()}`;
-        await store.storeBridgedMessage(fixture.message1.id, matrixMessageEventId);
+        await store.storeBridgedMessage(fixture.message1.id, matrixRoomId, matrixMessageEventId);
 
         await gitterBridge.onDataChange({
           url: `/rooms/${fixture.troupe1.id}/chatMessages`,
@@ -322,7 +320,8 @@ describe('gitter-bridge', () => {
 
         // We purposely do not associate the bridged message. We are testing that the
         // edit is ignored if there is no association in the database.
-        //await store.storeBridgedMessage(fixture.message1.id, matrixMessageEventId);
+        //const matrixRoomId = `!${fixtureLoader.generateGithubId()}:localhost`;
+        //await store.storeBridgedMessage(fixture.message1.id, matrixRoomId, matrixMessageEventId);
 
         await gitterBridge.onDataChange({
           url: `/rooms/${fixture.troupe1.id}/chatMessages`,
@@ -341,8 +340,13 @@ describe('gitter-bridge', () => {
           strategy
         );
 
+        const matrixRoomId = `!${fixtureLoader.generateGithubId()}:localhost`;
         const matrixMessageEventId = `$${fixtureLoader.generateGithubId()}`;
-        await store.storeBridgedMessage(fixture.messageFromVirtualUser1.id, matrixMessageEventId);
+        await store.storeBridgedMessage(
+          fixture.messageFromVirtualUser1.id,
+          matrixRoomId,
+          matrixMessageEventId
+        );
 
         await gitterBridge.onDataChange({
           url: `/rooms/${fixture.troupe1.id}/chatMessages`,
@@ -361,8 +365,13 @@ describe('gitter-bridge', () => {
           strategy
         );
 
+        const matrixRoomId = `!${fixtureLoader.generateGithubId()}:localhost`;
         const matrixMessageEventId = `$${fixtureLoader.generateGithubId()}`;
-        await store.storeBridgedMessage(fixture.messagePrivate1.id, matrixMessageEventId);
+        await store.storeBridgedMessage(
+          fixture.messagePrivate1.id,
+          matrixRoomId,
+          matrixMessageEventId
+        );
 
         await gitterBridge.onDataChange({
           url: `/rooms/${fixture.troupePrivate1.id}/chatMessages`,
@@ -401,8 +410,9 @@ describe('gitter-bridge', () => {
       });
 
       it('remove message gets sent off to Matrix', async () => {
+        const matrixRoomId = `!${fixtureLoader.generateGithubId()}:localhost`;
         const matrixMessageEventId = `$${fixtureLoader.generateGithubId()}`;
-        await store.storeBridgedMessage(fixture.message1.id, matrixMessageEventId);
+        await store.storeBridgedMessage(fixture.message1.id, matrixRoomId, matrixMessageEventId);
 
         await gitterBridge.onDataChange({
           url: `/rooms/${fixture.troupe1.id}/chatMessages`,
@@ -424,7 +434,8 @@ describe('gitter-bridge', () => {
       it('non-bridged message that gets removed is ignored', async () => {
         // We purposely do not associate bridged message. We are testing that the
         // remove is ignored if no association in the database.
-        //await store.storeBridgedMessage(fixture.message1.id, matrixMessageEventId);
+        //const matrixRoomId = `!${fixtureLoader.generateGithubId()}:localhost`;
+        //await store.storeBridgedMessage(fixture.message1.id, matrixRoomId, matrixMessageEventId);
 
         await gitterBridge.onDataChange({
           url: `/rooms/${fixture.troupe1.id}/chatMessages`,
@@ -437,8 +448,13 @@ describe('gitter-bridge', () => {
       });
 
       it('private room is not bridged', async () => {
+        const matrixRoomId = `!${fixtureLoader.generateGithubId()}:localhost`;
         const matrixMessageEventId = `$${fixtureLoader.generateGithubId()}`;
-        await store.storeBridgedMessage(fixture.messagePrivate1.id, matrixMessageEventId);
+        await store.storeBridgedMessage(
+          fixture.messagePrivate1.id,
+          matrixRoomId,
+          matrixMessageEventId
+        );
 
         await gitterBridge.onDataChange({
           url: `/rooms/${fixture.troupePrivate1.id}/chatMessages`,
@@ -456,8 +472,9 @@ describe('gitter-bridge', () => {
           throw new Error('Fake error and failed to fetch event');
         };
 
+        const matrixRoomId = `!${fixtureLoader.generateGithubId()}:localhost`;
         const matrixMessageEventId = `$${fixtureLoader.generateGithubId()}`;
-        await store.storeBridgedMessage(fixture.message1.id, matrixMessageEventId);
+        await store.storeBridgedMessage(fixture.message1.id, matrixRoomId, matrixMessageEventId);
 
         await gitterBridge.onDataChange({
           url: `/rooms/${fixture.troupe1.id}/chatMessages`,